Añadimos nuevas funcionalidades al proyecto y swagger para la documetación y pruebas

+++ src/main/java/demo/EasyDevel/demo/repository/DemoRepository.java
@@ -3,6 +3,6 @@ package demo.EasyDevel.demo.repository;
 import demo.EasyDevel.demo.model.Section;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DemoRepository extends JpaRepository<Section, Long> {
+public interface DemoRepository extends JpaRepository<Section, Integer> {
 
 }
